fix: resolve all mypy type errors across services, API, and utils layers (tasks 1.2-1.3)
Fix 84 genuine type errors to achieve clean mypy pass:
- Add Optional annotations for implicit Optional params (13 files)
- Add class-level type annotations to singleton services (6 classes)
- Fix return types on MediaService.add_movie/add_series/add_music
- Fix aiohttp timeout types in HealthService (int -> ClientTimeout)
- Fix BaseException vs Exception in asyncio.gather results
- Add missing return statement in BaseApiClient._make_request
- Fix ValidationError.message -> str(e) access

Result: mypy src/ passes clean (0 errors), 1869 tests pass, flake8 clean.

+++ src/utils/helpers.py
@@ -82,7 +82,7 @@ def save_chat_id(chat_id: int, chat_name: Optional[str] = None):
         f.write(entry)
 
 
-def format_bytes(size: int) -> str:
+def format_bytes(size: float) -> str:
     """Format bytes to human readable string"""
     for unit in ['B', 'KB', 'MB', 'GB', 'TB']:
         if size < 1024:
